Enable MkDocs publishing to GitHub Pages
This technique relies on using a Travis-encrypted GitHub Deploy Key,
which only has access to this repository. Publishing will only occur
when the build is not for a Pull Request.

Travis starts this build with a single-branch clone from the HTTPS URL,
which is what drives the need to alter the origin URL and configure the
GitHub Pages publishing branch.

There is some discussion on how to better automate this process in
MkDocs here:

https://github.com/mkdocs/mkdocs/issues/632

Note that there are two approaches that can be used:

* GitHub Personal Access Token
* GitHub Deploy Key

The Personal Access Token method means that whomever has your token can
act as you, with with whatever access you have been granted to repos,
subject to the configured OAuth scopes.

The Deploy Key method allows you to restrict access to a single
repository, which is ideal for this use case.