add mkdocs-atlas-formatting-plugin
This is a custom MkDocs plugin, which parses directives in Markdown source
and replaces them with either Atlas graphs or formatted Atlas expressions.
These illustrations help to make Atlas usage examples in the documentation
more concrete.

A full description of the directives and how to use them can be found in the
README for the plugin.