add atlas-uri directive and hilite option
The atlas-uri directive is similar to the atlas-stacklang directive, except that
it shows the full URI, so that URL parameters can be demonstrated.

A hilite option has been added to the atlas-stacklang and atlas-uri directives,
which adds a light yellow background color span to a specified chunk of text. Any
equals signs in the text selection should be percent encoded (%3d), so as not to
break the option parsing.

Additional tests were added around block option processing, to ensure that they
remain robust in the face of bad input.