shutdown atlas standalone webserver after the build
Since the port detection and fail-to-start exception was added, to assist with
correctness when developing locally with `mkdocs serve`, the Travis build has
been failing, because the `mkdocs build` step starts the webserver and does not
stop it. Thus, when the `mkdocs gh-deploy` step runs to deploy the site, it fails.

This change adds a check to the post_build event, so that if the `mkdocs build`
command is running, it will shutdown the webserver before proceeding.

Additional details have also been added to setup.py.