add atlas-link-checker cli tool (#64)
As we build up the new Atlas Docs site, we want to be able to validate that the
links in the site are correct. However, we do not want this check to occur with
the site build process, because we do not want transient connectivity errors to
fail the build.

This CLI tool is intended to be run as a one-off after the site has been built,
so that you can verify the status of the links embedded within the pages.