update ASL reference for :and (#67)
Replaces the old stack effect diagram with something that
should be more approachable for users new to stack based
languages.

Update the examples to follow the internal ASL course. The
outline for the ASL tutorial has been filled in to allow
cross-linking, but the content has not been moved over yet.